feat(pins): add apply button and share chunks cache with diffs.
