+++ src/events/index.ts
@@ -17,6 +17,10 @@ import type { FIMDebugState } from "../hooks";
 export { updateConfig, type Config } from "../features/Config/configSlice";
 export { type FileInfo, setFileInfo } from "../features/Chat/activeFile";
 export {
+  setOpenFiles,
+  type OpenFilesState,
+} from "../features/OpenFiles/openFilesSlice";
+export {
   type Snippet,
   setSelectedSnippet,
 } from "../features/Chat/selectedSnippet";
