+++ src/services/refact/diffs.ts
@@ -1,6 +1,11 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import { DIFF_STATE_URL, DIFF_APPLY_URL, DIFF_PREVIEW_URL } from "./consts";
-import { DiffChunk, isDiffErrorResponseData } from "./types";
+import {
+  DIFF_STATE_URL,
+  DIFF_APPLY_URL,
+  DIFF_PREVIEW_URL,
+  PATCH_URL,
+} from "./consts";
+import { DiffChunk, isDiffChunk, isDiffErrorResponseData } from "./types";
 import { RootState } from "../../app/store";
 
 export type DiffAppliedStateArgs = {
@@ -49,6 +54,78 @@ export type DiffApplyManyArgs = {
   toApply: boolean;
 }[];
 
+type PatchState = {
+  chunk_id: number;
+  applied: boolean;
+  can_unapply: boolean;
+  success: boolean;
+  detail: null | string;
+};
+
+function isPatchState(json: unknown): json is PatchState {
+  if (!json || typeof json !== "object") return false;
+  if (!("chunk_id" in json)) return false;
+  if (typeof json.chunk_id !== "number") return false;
+  if (!("applied" in json)) return false;
+  if (typeof json.applied !== "boolean") return false;
+  if (!("can_unapply" in json)) return false;
+  if (typeof json.can_unapply !== "boolean") return false;
+  if (!("success" in json)) return false;
+  if (typeof json.success !== "boolean") return false;
+  return true;
+}
+
+type PatchResult = {
+  file_text: string;
+  file_name_edit: string | null;
+  file_name_delete: string | null;
+  file_name_add: string | null;
+};
+
+function isPatchResult(json: unknown): json is PatchResult {
+  if (!json || typeof json !== "object") return false;
+  if (!("file_text" in json)) return false;
+  if (typeof json.file_text !== "string") return false;
+  if (!("file_name_edit" in json)) return false;
+  if (typeof json.file_name_edit !== "string" && json.file_name_edit !== null)
+    return false;
+  if (!("file_name_delete" in json)) return false;
+  if (
+    typeof json.file_name_delete !== "string" &&
+    json.file_name_delete !== null
+  )
+    return false;
+  if (!("file_name_add" in json)) return false;
+  if (typeof json.file_name_add !== "string" && json.file_name_add !== null)
+    return false;
+  return true;
+}
+
+type PatchResponse = {
+  state: PatchState[];
+  results: PatchResult[];
+  chunks: DiffChunk[];
+};
+
+function isPatchResponse(json: unknown): json is PatchResponse {
+  if (!json || typeof json !== "object") return false;
+  if (!("state" in json)) return false;
+  if (!Array.isArray(json.state)) return false;
+  if (!json.state.every(isPatchState)) return false;
+  if (!("results" in json)) return false;
+  if (!Array.isArray(json.results)) return false;
+  if (!json.results.every(isPatchResult)) return false;
+  if (!("chunks" in json)) return false;
+  if (!Array.isArray(json.chunks)) return false;
+  if (!json.chunks.every(isDiffChunk)) return false;
+  return true;
+}
+
+type PatchRequest = {
+  pin: string;
+  markdown: string;
+};
+
 // function cacheIdForChunk(chunk: DiffChunk) {
 //   return `${chunk.file_name}-${chunk.line1}-${chunk.line2}`;
 // }
@@ -66,7 +143,7 @@ export const diffApi = createApi({
       return headers;
     },
   }),
-  tagTypes: ["DIFF_STATE", "DIFF_PREVIEW"],
+  tagTypes: ["DIFF_STATE", "DIFF_PREVIEW", "DIFF_PATCH"],
   endpoints: (builder) => ({
     diffState: builder.query<DiffStateResponse[], DiffAppliedStateArgs>({
       queryFn: async (args, api, _extraOptions, baseQuery) => {
@@ -139,7 +216,7 @@ export const diffApi = createApi({
       },
       invalidatesTags: (_result, _error, _args) => {
         // TODO: this could be more efficient
-        return [{ type: "DIFF_STATE" }];
+        return [{ type: "DIFF_STATE" }, { type: "DIFF_PATCH" }];
       },
     }),
 
@@ -172,7 +249,47 @@ export const diffApi = createApi({
       //   });
       // },
     }),
+
+    patchSingleFileFromTicket: builder.query<PatchResponse, PatchRequest>({
+      providesTags: ["DIFF_PATCH"],
+      async queryFn(args, api, _extraOptions, baseQuery) {
+        const state = api.getState() as RootState;
+        const port = state.config.lspPort as unknown as number;
+        const url = `http://127.0.0.1:${port}${PATCH_URL}`;
+
+        const ticket = args.pin.split(" ")[1] ?? "";
+        const messages = [
+          { role: "assistant", content: args.pin + "\n" + args.markdown },
+        ];
+
+        const result = await baseQuery({
+          url,
+          credentials: "same-origin",
+          redirect: "follow",
+          method: "POST",
+          body: {
+            messages,
+            ticket_ids: [ticket],
+          },
+        });
+
+        if (result.error) return { error: result.error };
+
+        if (!isPatchResponse(result.data)) {
+          return {
+            error: {
+              status: "CUSTOM_ERROR",
+              error: "Failed to parse patch response",
+              data: result.data,
+            },
+          };
+        }
+
+        return { data: result.data };
+      },
+    }),
   }),
+
   refetchOnMountOrArgChange: true,
 });
